CIFAR10 loading
The training functions expects dataset to be available as a set of directories with images.  We have included the code to execute dataset extraction. To do so, download Cifar 10 as a zip file https://www.kaggle.com/datasets/swaroopkml/cifar10-pngs-in-folders?resource=download and store it in the same directory as your code. Doing so will create a folder: cifar10/cifar10. We store the inner folder directly into your root directory for the project.